Got rid of most subclasses and cleaned up code
The OfficeClock now multiply inherits ButtonManager, WUnderground and BrightnessManager.
This avoids the need for keeping a reference to OfficeClock in all the subclasses.

Moved updateDisplay to a setTime method of ClockDisplay.

There is still a WifiManager subclass because it uses a static method, so we  need
a thunk to get to OfficeClock. I plan on adding a virtual method to the WifiManager
project to solve this later. But I've also come across PersWifiManager, which is
an offshoot that supposedly allows async network access and more web server
capabilities