Switched to using a StateMachine
Got rid of all the individual flags and replaced with a StateMachine.
Still need a loop to handle the call to wUnderground since that
needs to happen in the main thread.

Everything works and now there are more descriptive messages output
for all the states. Can get into Setup, but nothing happens there.